个人主页没传id

+++ lib/page/profile/profile_page.dart
@@ -41,7 +41,7 @@ class ProfileState extends State<ProfilePage> with AutomaticKeepAliveClientMixin
   int followerCount;
   int feedCount;
   int trainingSeconds;
-  int weight;
+  double weight;
   int albumNum;
 
   @override
@@ -139,7 +139,7 @@ class ProfileState extends State<ProfilePage> with AutomaticKeepAliveClientMixin
                   children: [
                     _secondData(Icons.access_alarms_sharp, trainingSeconds, "训练记录", height, width),
                     Expanded(child: Container()),
-                    _secondData(Icons.access_alarms_sharp, weight, "体重记录", height, width),
+                    _secondData(Icons.access_alarms_sharp, weight.toInt(), "体重记录", height, width),
                     Expanded(child: Container()),
                     _secondData(Icons.access_alarms_sharp, albumNum, "健身相册", height, width),
                   ],
